:recycle: moved constants to @static/leaderboard
