+++ src/pages/user/leaderboard.tsx
@@ -1,6 +1,6 @@
 import UserLeaderboardComponent from "@components/pages/user/leaderboard";
 import { LeaderboardFilterProvider } from "@hooks/useLeaderboardFilter";
-import { LEADERBOARD_FILTERS } from "@static/constants";
+import { LEADERBOARD_FILTERS } from "@static/leaderboard";
 import React from "react";
 
 function UserLeaderboardPage({ initialFilterParams }) {
